requirements.txt
kaggle==1.5.12
numpy==1.22.3
Pillow==9.0.1
matplotlib==3.5.1
albumentations==1.1.0 #look for using nn.transforms instead